+++ src/expression/index.ts
@@ -61,6 +61,7 @@ export type GlobalProperties = Readonly<{
     lineProgress?: number;
     isSupportedScript?: (_: string) => boolean;
     accumulated?: Value;
+    globalState?: Record<string, any>;
 }>;
 
 export class StyleExpression {